Adds an abstract class that injects a map fragment
